Switch to yield for returning samples during generation

+++ scripts/optimization/pixelvae.py
@@ -167,17 +167,17 @@ class PixelVAE:
         self.model.to(device)
 
     def run_plain(self, samples):
-        predicts = self.model.forward(samples.to(self.device))
+        predicts = self.model.forward(samples.to(self.device).to(torch.float32))
         result = net_output_to_color8(self.binning, predicts)
         return result
 
     def run_paletted(self, samples, palette):
-        predicts = self.model.forward(samples.to(self.device))
+        predicts = self.model.forward(samples.to(self.device).to(torch.float32))
         result = net_output_to_color8_pal(self.binning, predicts, palette)
         return result
 
     def run_cluster(self, samples, threshold=0.001, rand_seed=1, select='local8', wrap_x=False, wrap_y=False):
-        predicts = self.model.forward(samples.to(self.device)).to(torch.float32)
+        predicts = self.model.forward(samples.to(self.device).to(torch.float32))
 
         # compute softmax per channel, split channels
         compute_softmax(self.binning, predicts)
